[vivihuang][TG] *67 fix: use dataframe to save space for chat.

+++ src/action/actions/tb_guru/wcs_data_query.py
@@ -1,7 +1,6 @@
 import os
 import shutil
 import uuid
-from typing import Union, Any
 
 import pandas as pd
 from gluon_meson_sdk.models.abstract_models.chat_message_preparation import ChatMessagePreparation
@@ -16,7 +15,7 @@ from action.base import (
     ActionResponse,
     UploadFileContentType,
 )
-from third_system.search_entity import SearchItem, SearchParam
+from third_system.search_entity import SearchParam
 from third_system.unified_search import UnifiedSearch
 from utils.ppt_helper import plot_graph, ppt_generation
 
@@ -46,32 +45,29 @@ class WcsDataQuery(Action):
     def get_name(self) -> str:
         return "wcs_data_query"
 
-    def _generate_ppt(
-        self, current_company_data: list[SearchItem], all_companies_data: list[SearchItem], insight: str
-    ) -> str:
-        files_dir = f"{self.tmp_file_dir}/{str(uuid.uuid4())}"
-        os.makedirs(files_dir, exist_ok=True)
-        image_paths = [f"{files_dir}/image1.png", f"{files_dir}/image2.png"]
-        ppt_path = f"{files_dir}/ppt.pptx"
-        df_current = pd.DataFrame([w.model_dump() for w in current_company_data])
-        df_all = pd.DataFrame([w.model_dump() for w in all_companies_data])
-        company_name = current_company_data[0].company
-        plot_graph(df_current, f"{company_name} – Working Capital Metrics Trend", "days", image_paths[0])
-        plot_graph(df_all, f"Peer Comparison ({all_companies_data[0].days})", "company", image_paths[1])
-        df_all = df_all.drop(columns=["days", "meta__score", "meta__reference", "id"])
-        ppt_generation(
-            df_all, llm_insight=insight, company_name=company_name, image_paths=image_paths, output_path=ppt_path
-        )
-        files = [
-            ("files", ("tb_guru_ppt.pptx", open(ppt_path, "rb"), UploadFileContentType.PPTX)),
-        ]
-        links = self.unified_search.upload_file_to_minio(files)
-        shutil.rmtree(files_dir)
-        return links[0] if links else ""
+    def _generate_ppt(self, df_current: pd.DataFrame, df_all: pd.DataFrame, insight: str) -> str:
+        company_name = df_current.iloc[0]["company"] if "company" in df_current.iloc[0] else ""
+        latest_days = df_all.iloc[0]["days"] if "days" in df_all.iloc[0] else ""
+        if company_name and latest_days:
+            files_dir = f"{self.tmp_file_dir}/{str(uuid.uuid4())}"
+            os.makedirs(files_dir, exist_ok=True)
+            image_paths = [f"{files_dir}/image1.png", f"{files_dir}/image2.png"]
+            ppt_path = f"{files_dir}/ppt.pptx"
+            plot_graph(df_current, f"{company_name} – Working Capital Metrics Trend", "days", image_paths[0])
+            plot_graph(df_all, f"Peer Comparison ({latest_days})", "company", image_paths[1])
+            df_all = df_all.drop(columns=["days"])
+            ppt_generation(
+                df_all, llm_insight=insight, company_name=company_name, image_paths=image_paths, output_path=ppt_path
+            )
+            files = [
+                ("files", ("tb_guru_ppt.pptx", open(ppt_path, "rb"), UploadFileContentType.PPTX)),
+            ]
+            links = self.unified_search.upload_file_to_minio(files)
+            shutil.rmtree(files_dir)
+            return links[0] if links else ""
+        return ""
 
-    async def _search(
-        self, entity_dict, session_id
-    ) -> tuple[Union[list[SearchItem], list[Any]], Union[list[SearchItem], list[Any]]]:
+    async def _search(self, entity_dict, session_id) -> tuple[pd.DataFrame, pd.DataFrame]:
         query_list = [
             f"""Query WCS data with days as {entity_dict["days"]}:"""
             if "days" in entity_dict
@@ -83,9 +79,12 @@ class WcsDataQuery(Action):
         items = [s[0].items if s else [] for s in search_res]
         all_companies_data, current_company_data = items[0], items[1] if len(items) > 1 else []
         latest_period = all_companies_data[0].days if all_companies_data else ""
-        return list(
-            filter(lambda s: s.days == latest_period, all_companies_data)
-        ) if latest_period else [], current_company_data
+        latest_all_data = list(filter(lambda s: s.days == latest_period, all_companies_data)) if latest_period else []
+        df_current = pd.DataFrame([w.model_dump() for w in current_company_data])
+        df_all = pd.DataFrame([w.model_dump() for w in latest_all_data])
+        df_current = df_current.drop(columns=["meta__score", "meta__reference", "id"])
+        df_all = df_all.drop(columns=["meta__score", "meta__reference", "id"])
+        return df_all, df_current
 
     async def run(self, context) -> ActionResponse:
         chat_model = self.scenario_model_registry.get_model(self.scenario_model, context.conversation.session_id)
@@ -94,7 +93,9 @@ class WcsDataQuery(Action):
         entity_dict = context.conversation.get_simplified_entities()
         is_ppt_output = entity_dict["is_ppt_output"] if "is_ppt_output" in entity_dict else False
 
-        all_companies_data, current_company_data = await self._search(entity_dict, context.conversation.session_id)
+        df_all_companies_data, df_current_company_data = await self._search(
+            entity_dict, context.conversation.session_id
+        )
 
         chat_message_preparation = ChatMessagePreparation()
         chat_message_preparation.add_message(
@@ -104,8 +105,9 @@ class WcsDataQuery(Action):
         )
         chat_message_preparation.add_message(
             "assistant",
-            """## WCS data\n{{wcs_data}}""",
-            wcs_data="\n".join([item.model_dump_json() for item in current_company_data + all_companies_data]),
+            """## WCS data are extracted already\n{{wcs_data}}""",
+            wcs_data=df_all_companies_data.to_json(orient="records")
+            + df_current_company_data.to_json(orient="records"),
         )
         chat_message_preparation.log(logger)
 
@@ -114,11 +116,11 @@ class WcsDataQuery(Action):
         ).response
         logger.info(f"chat result: {result}")
 
-        if is_ppt_output and current_company_data and all_companies_data:
-            ppt_link = self._generate_ppt(current_company_data, all_companies_data, result)
+        if is_ppt_output and not df_current_company_data.empty and not df_all_companies_data.empty:
+            ppt_link = self._generate_ppt(df_current_company_data, df_all_companies_data, result)
 
             chat_message_preparation = ChatMessagePreparation()
-            chat_message_preparation.add_message("assistant", ppt_prompt, ppt_link=ppt_link)
+            chat_message_preparation.add_message("system", ppt_prompt, ppt_link=ppt_link)
             chat_message_preparation.log(logger)
 
             result = chat_model.chat(
